Handle stale promises in attemptConsume
Previously, we had special handling for stale promises in
consumeStalePromises. That function manually reached into the DB and
removed them directly.

However, removing stale promises is really a special case of consuming
them - in both cases the intent is "this promise is no longer needed
and can be removed from everywhere we keep track of it" - the only
difference between stale and non-stale promises is that it is not
possible to perform the UI action in the stale case (although you may
want to perform some other action, such as showing an error popup).

Therefore: generalise attemptConsume to handle both stale and non-stale
promises - only calling the AnyPromise if the promise is stale, but
removing it from the DB in either case. As an extra bonus, remove the
removeAllPromises: method which was only used here.