Include all monalxmpp headers in target
All the public headers of monalxmpp (the culmination of those in
monalxmpp.h, and those imported indirectly from monalxmpp.h) need to be
both included in the monalxmpp target, and publicly available.

Without this, on importing the briding header from outside monalxmpp
(which we need to do from the monal app delegate) the imports fail at
compile time.