Skip tests that always fail
Traditionally in Monal, tests have not been automatically run against
PRs. We are thinking of changing this and running tests against certain
PRs. As a result, we want to skip tests that always fail. It so happens
that these are all UI tests.

I have skipped the individual tests rather than excluding (or deleting)
the whole of the whole UI tests target, since we will likely want to
reintroduce UI tests in the future.